Support debug_ranges lookup when matching address for inline functions.
Summary:
Range lists are contained in a separate object file section called .debug_ranges. A range list is
indicated by a DW_AT_ranges attribute whose value is represented as an offset from the
beginning of the .debug_ranges section to the beginning of the range list.

Each entry in a range list is either:
- a range list entry,
- a base address selection entry, or
- an end of list entry.

Reviewed By: luciang

Differential Revision: D19438098

fbshipit-source-id: 12f1d5c60067b0674c44f35ef2731781eb93f1d5