fix compilation issues on MSVC
Summary:
MSVC seems to have a bug when handling private classes that
are both part of the inheritence hierarchy (private superclass of a
superclass) and available as a template parameter, incorrectly claiming
that the class is inaccessible due to private inheritance. This diff
works around the issue. It also adds an explicit case to avoid a narrowing
warning generated by MSVC.

(Note: this ignores all push blocking failures!)

Reviewed By: yfeldblum, mjoras

Differential Revision: D19561643

fbshipit-source-id: 494d1611ce662a0eb6c441c64479d3c5d2bbd5b6