fix missing eraseInto overload for key with permissive constructor
Summary:
If a key type has a templated constructor that can implicitly
accept const_iterator, then a call to erase or eraseInto with a non-const
iterator results in an ambiguous overload. This diff adds an extra
overload of eraseInto to avoid the ambiguity, as well as adding unit
tests that cover the rare corner case.

Reviewed By: yfeldblum

Differential Revision: D19809069

fbshipit-source-id: 499849da2d6ad64070b13eb0a9ed86a3d49ad9b0