work around compiler bug on some platforms
Summary: g++ aggressively inlines F14VectorMap::operator[] calls from some testing loops with the loop index as a key. On platforms that have SSE2 but not SSE4.2 we use a multiplication and xor based bit mixer, which undergoes a strength reduction in optimized builds. Somehow the resulting code miscalculates the result (no minimal repro yet), which causes the tests to persistently fail. This diff blocks the compiler optimization, allowing the tests to pass.

Reviewed By: yfeldblum

Differential Revision: D19910180

fbshipit-source-id: 6cccba67645481061f923bd7b3e681df573c4039