Work around spurious -Werror=array-bounds error in gcc-9.3.
Summary:
gcc-9.3 complains about folly::Format:

```
array subscript -1 is outside array bounds of ‘char [69]’
```

But the warning is incorrect; my guess is that it is confused because
it doesn't realize that higher-level logic guarantees that integer
formatting will run out of digits before it runs out of buffer.

Changing some code from `array + x` to `&array[x]` works around it.

Reviewed By: pixelb

Differential Revision: D20694399

fbshipit-source-id: 1176858054cdec9e415de4829371647958955103