Let small_vector pop_back a non-assignable type
Summary: Relying on `erase()`, which must be able to move elements when erasing in front of other valid elements, causes compilation failures when trying to erase non-assignable elements at the end of the vector even though no elements actually need to be moved (and thus assigned) in that case. Here we help the compiler out and extract the destruction/resizing logic away from the moving logic.

Reviewed By: yfeldblum, ot, mkatsevVR

Differential Revision: D20718712

fbshipit-source-id: 51c4e6636bc33577494c0c10e95da5f92b58d4b6