Disable TSAN instrumentation across StaticMetaBase::onThreadExit().
Summary:
Disable TSAN instrumentation across StaticMetaBase::onThreadExit() to avoid a shortcoming in TSAN on Linux that produces sprurious failures.

This was recently worked around by ignoring TSAN failures in ld-linux's _dl_deallocate_tsd() (D20597534), but that turns out to be insufficient, because other TSAN failures can occur on the TLS area after the dynamic linker has called free() on it.  This diff backs out the change from D20597534, and works around the problem in a more direct & complete way.

Reviewed By: yfeldblum

Differential Revision: D20671620

fbshipit-source-id: 05a74e709b620391f18467c7a5e4d14dcf9f4a1e