windows: default to c++17
Summary:
While compiling EdenFS on Windows, for every single file being compiled,
cl.exe complains that 2 incompatible /std option are being passed in on the
command line, a c++latest, and a c++17. The former is added by this file,
while the later is added by cmake via CMAKE_CXX_STANDARD.

I'm not exactly sure why the default is set to c++latest, but it's guaranteed
to always force cl.exe to generate a warning, as it simply cannot be set via
CMAKE_CXX_STANDARD.

A better fix to this would be to set CMAKE_CXX_STANDARD to 17, but since
folly still supports cmake 3.0, that didn't know about c++17, this can't
be done.

Reviewed By: yfeldblum

Differential Revision: D21392360

fbshipit-source-id: fc0e29ed89adf0a2c0b0193e7bd5f033463ed376