define FOLLY_XLOG_STRIP_PREFIXES correctly for all targets
Summary:
The `FOLLY_XLOG_STRIP_PREFIXES` macro tells folly::logging how to strip off
the absolute portion of the path when automatically computing log category
names from the source file name.

Previously this was set in FollyCompilerUnix.cmake, which caused it to only be
applied on Linux/Mac platforms, and only for targets that explicitly used
`apply_folly_compile_options_to_target()`.

This changes the code to apply this definition with a directory property, so
it affects all targets under the top-level directory, and it is also applied
on Windows.

Reviewed By: wez

Differential Revision: D21333014

fbshipit-source-id: 9256e14ab4fd03d19f28957e7ca298c930a22b18