make it possible to use folly logging with exceptions disabled
Summary:
This updates the logging code so that it is possible to include `folly/xlog.h`
and write log statements in code that is compiled with `-fno-exceptions`

The code for parsing log configuration strings does still use exceptions in
some of its config parsing logic.  It shouldn't be too difficult to disable
that in a subsequent diff, but for now this change at least makes it possible
for calling code to write log messages.

Reviewed By: yfeldblum

Differential Revision: D21630914

fbshipit-source-id: ce9ed422f40066e57433a85182e1a8fc6d8628db