Remove collectXSemiFuture
Summary:
Migration from Future-returning executor-erasing collectX forms to
SemiFuture-returning forms, that are less risky in particular with coroutines.

Earlier changes added collectXSemiFuture and collectXUnsafe as a migration
path. We then migrated collectX callsites to collectXSemiFuture or
collectXUnsafe and switched the implementation of collectX to the SemiFuture
form.

This diff removes collectXSemiFuture completely.

Reviewed By: yfeldblum, mpark

Differential Revision: D21631886

fbshipit-source-id: c7b296d51f30c8b51ab0c7d06b1ce62c8cdd28d5