add a method to access a threadlocal without creating an object
Summary:
It may be useful to see if a threadlocal has been created on a thread before accessing it.

One way this could be useful: the data is only relevant to a few threads but will be accessed by many. This would allow you to save the memory on the non-relevant threads.

Differential Revision: D21731260

fbshipit-source-id: 0b1835b454f9b0c457f722ab38f84f89aca383fb