don't use clock_gettime fallback on modern macOS
Summary:
There were two problems with the clock_gettime fallback logic on
macOS. One is that it used undefined macros if AvailabilityMacros.h
wasn't included. Another is that it compiled the fallback when the
minimum iPhone OS version (as when building on macOS) was not set, and
vice versa.

Fix that.

Differential Revision: D21757362

fbshipit-source-id: 99fd17260cb25ed07c39ba79466e835f08a1fc7c